requirements.txt
flask==2.3.3
openai
python-dotenv

